refactor(chat): extract MessageEditingDelegate and ChatRequestBuilder
Move the message-mutation send paths (edit user/assistant, regenerate,
continue) and the inline edit-field UI state into MessageEditingDelegate.
Extract the per-send request pieces shared by every send path
(buildEphemeralAgent + currentDispatch) into a pure ChatRequestBuilder used by
both the new-message path and the editing delegate.

ChatViewModel is now a thin orchestrator wiring the delegates; the
send-readiness gate, message-text extraction, and Room observer stay in it.
No behavior change.
