+++ feature/chat/src/commonMain/kotlin/com/garfiec/librechat/feature/chat/viewmodel/delegate/MessageEditingDelegate.kt
@@ -1,0 +1,274 @@
+package com.garfiec.librechat.feature.chat.viewmodel.delegate
+
+import co.touchlab.kermit.Logger
+import com.garfiec.librechat.core.common.EndpointConstants
+import com.garfiec.librechat.core.data.repository.ChatRepository
+import com.garfiec.librechat.core.data.repository.MessageRepository
+import com.garfiec.librechat.core.model.Message
+import com.garfiec.librechat.feature.chat.util.buildActiveMessagePath
+import com.garfiec.librechat.feature.chat.viewmodel.ChatRequestBuilder
+import com.garfiec.librechat.feature.chat.viewmodel.ChatStateHandle
+import kotlinx.coroutines.launch
+import kotlin.time.Clock
+import kotlin.uuid.ExperimentalUuidApi
+import kotlin.uuid.Uuid
+
+/**
+ * Owns the message-mutation send paths — edit (user / assistant), regenerate, and continue —
+ * plus the inline edit-field UI state. Each path optimistically reshapes the tree (a new
+ * sibling or a stream anchor via [MessageTreeDelegate]) and resubmits through
+ * [StreamingManagerDelegate], reusing [ChatRequestBuilder] for the shared request pieces.
+ *
+ * The send-readiness gate and message-text extraction live in `ChatViewModel` (shared with
+ * the new-message send path and TTS); they're injected as [runWhenSendReady] / [getMessageText].
+ */
+class MessageEditingDelegate(
+    private val stateHandle: ChatStateHandle,
+    private val chatRepository: ChatRepository,
+    private val messageRepository: MessageRepository,
+    private val treeDelegate: MessageTreeDelegate,
+    private val streamingManager: StreamingManagerDelegate,
+    private val requestBuilder: ChatRequestBuilder,
+    private val getMessageText: (String) -> String,
+    private val runWhenSendReady: (action: () -> Unit) -> Unit,
+) {
+
+    fun editMessage(messageId: String, newText: String) {
+        if (newText.isBlank() || stateHandle.state.isStreaming) return
+
+        val originalMessage = stateHandle.state.messages.find { it.messageId == messageId } ?: return
+
+        runWhenSendReady {
+            if (originalMessage.isCreatedByUser) {
+                editUserMessage(originalMessage, newText)
+            } else {
+                editAiMessage(originalMessage)
+            }
+        }
+    }
+
+    @OptIn(ExperimentalUuidApi::class)
+    private fun editUserMessage(originalMessage: Message, newText: String) {
+        val parentMessageId = originalMessage.parentMessageId
+
+        // Optimistically insert the edited text as a new sibling of the original user
+        // message and anchor the stream to it, so the new message — with the response
+        // streaming below it — replaces the old branch in place. The anchor truncates
+        // the active path here (see buildActiveMessagePath), mirroring the web client's
+        // `currentMsg + initialResponse` placeholder insert. The optimistic id is sent
+        // as `userMessageId` so the server adopts it; that lets the message reconcile by
+        // id on Final — via loadConversation() for normal chats, or in-memory via
+        // mergeFinalMessagesInMemory for temp chats (which never touch Room).
+        val optimisticMessage = Message(
+            messageId = Uuid.random().toString(),
+            conversationId = stateHandle.state.conversationId ?: "",
+            parentMessageId = parentMessageId,
+            text = newText,
+            isCreatedByUser = true,
+            sender = "User",
+            createdAt = Clock.System.now().toString(),
+            files = originalMessage.files,
+        )
+        stateHandle.update {
+            val updatedMessages = messages + optimisticMessage
+            copy(
+                messages = updatedMessages,
+                displayMessages = buildActiveMessagePath(updatedMessages, activeBranches, optimisticMessage.messageId),
+            )
+        }
+
+        streamingManager.prepareForStreaming(isEdit = true)
+
+        val state = stateHandle.state
+        val isAgent = state.selectedEndpoint == EndpointConstants.AGENTS
+        val webSearchEnabled = state.modelParameters.webSearch
+        val ephemeralAgent = requestBuilder.buildEphemeralAgent()
+        Logger.d { "editUserMessage: webSearch=$webSearchEnabled, ephemeralAgent=$ephemeralAgent" }
+        val dispatch = requestBuilder.currentDispatch()
+        streamingManager.launchStream(
+            chatRepository.startChat(
+                text = newText,
+                conversationId = state.conversationId,
+                endpoint = state.selectedEndpoint,
+                endpointType = dispatch.endpointType,
+                key = dispatch.key,
+                modelDisplayLabel = dispatch.modelDisplayLabel,
+                model = state.selectedModel,
+                userMessageId = optimisticMessage.messageId,
+                parentMessageId = parentMessageId,
+                agentId = if (isAgent) state.selectedModel else null,
+                isEdited = true,
+                webSearch = webSearchEnabled,
+                ephemeralAgent = ephemeralAgent,
+                isTemporary = state.isTemporaryChat,
+            ),
+        )
+    }
+
+    private fun editAiMessage(aiMessage: Message) {
+        val parentUserMessage = stateHandle.state.messages.find {
+            it.messageId == aiMessage.parentMessageId
+        } ?: return
+
+        val conversationId = stateHandle.state.conversationId ?: return
+
+        // Editing an assistant message resubmits its parent user turn (isEdited +
+        // isRegenerate) — the same shape as regenerate, so anchor the stream to the
+        // parent user message and let the new response stream in below it, replacing
+        // the old one. The web client seeds the placeholder with the edited content
+        // for a transient preview; we don't (the regenerated server response is
+        // authoritative on Final either way).
+        treeDelegate.anchorStreamTo(parentUserMessage.messageId)
+        streamingManager.prepareForStreaming(isEdit = true)
+
+        val state = stateHandle.state
+        val isAgent = state.selectedEndpoint == EndpointConstants.AGENTS
+        val webSearchEnabled = state.modelParameters.webSearch
+        val ephemeralAgent = requestBuilder.buildEphemeralAgent()
+        Logger.d { "editAiMessage: webSearch=$webSearchEnabled, ephemeralAgent=$ephemeralAgent" }
+        val dispatch = requestBuilder.currentDispatch()
+        streamingManager.launchStream(
+            chatRepository.startChat(
+                text = parentUserMessage.text,
+                conversationId = conversationId,
+                endpoint = state.selectedEndpoint,
+                endpointType = dispatch.endpointType,
+                key = dispatch.key,
+                modelDisplayLabel = dispatch.modelDisplayLabel,
+                model = state.selectedModel,
+                parentMessageId = parentUserMessage.parentMessageId,
+                agentId = if (isAgent) state.selectedModel else null,
+                overrideParentMessageId = parentUserMessage.messageId,
+                isEdited = true,
+                isRegenerate = true,
+                webSearch = webSearchEnabled,
+                ephemeralAgent = ephemeralAgent,
+                isTemporary = state.isTemporaryChat,
+            ),
+        )
+    }
+
+    fun regenerateMessage(messageId: String) {
+        if (stateHandle.state.isStreaming) return
+
+        val aiMessage = stateHandle.state.messages.find { it.messageId == messageId } ?: return
+        if (aiMessage.isCreatedByUser) return
+
+        val parentUserMessage = stateHandle.state.messages.find {
+            it.messageId == aiMessage.parentMessageId
+        } ?: return
+
+        runWhenSendReady { regenerateMessageNow(parentUserMessage) }
+    }
+
+    private fun regenerateMessageNow(parentUserMessage: Message) {
+        treeDelegate.anchorStreamTo(parentUserMessage.messageId)
+        streamingManager.prepareForStreaming(isEdit = true)
+
+        val state = stateHandle.state
+        val isAgentRegen = state.selectedEndpoint == EndpointConstants.AGENTS
+        val webSearchEnabled = state.modelParameters.webSearch
+        val ephemeralAgent = requestBuilder.buildEphemeralAgent()
+        Logger.d { "regenerateMessage: webSearch=$webSearchEnabled, ephemeralAgent=$ephemeralAgent" }
+        val dispatch = requestBuilder.currentDispatch()
+        streamingManager.launchStream(
+            chatRepository.startChat(
+                text = parentUserMessage.text,
+                conversationId = state.conversationId,
+                endpoint = state.selectedEndpoint,
+                endpointType = dispatch.endpointType,
+                key = dispatch.key,
+                modelDisplayLabel = dispatch.modelDisplayLabel,
+                model = state.selectedModel,
+                parentMessageId = parentUserMessage.parentMessageId,
+                agentId = if (isAgentRegen) state.selectedModel else null,
+                overrideParentMessageId = parentUserMessage.messageId,
+                isRegenerate = true,
+                webSearch = webSearchEnabled,
+                ephemeralAgent = ephemeralAgent,
+                isTemporary = state.isTemporaryChat,
+            ),
+        )
+    }
+
+    fun continueGeneration() {
+        if (stateHandle.state.isStreaming) return
+        val lastAiMessage = stateHandle.state.displayMessages.lastOrNull {
+            !it.message.isCreatedByUser
+        } ?: return
+
+        val parentUserMessage = stateHandle.state.messages.find {
+            it.messageId == lastAiMessage.message.parentMessageId
+        } ?: return
+
+        runWhenSendReady { continueGenerationNow(lastAiMessage.message, parentUserMessage) }
+    }
+
+    private fun continueGenerationNow(lastAiMessage: Message, parentUserMessage: Message) {
+        streamingManager.prepareForStreaming(isEdit = true)
+
+        val state = stateHandle.state
+        val isAgentContinue = state.selectedEndpoint == EndpointConstants.AGENTS
+        val webSearchEnabled = state.modelParameters.webSearch
+        val ephemeralAgent = requestBuilder.buildEphemeralAgent()
+        Logger.d { "continueGeneration: webSearch=$webSearchEnabled, ephemeralAgent=$ephemeralAgent" }
+        val dispatch = requestBuilder.currentDispatch()
+        streamingManager.launchStream(
+            chatRepository.startChat(
+                text = parentUserMessage.text,
+                conversationId = state.conversationId,
+                endpoint = state.selectedEndpoint,
+                endpointType = dispatch.endpointType,
+                key = dispatch.key,
+                modelDisplayLabel = dispatch.modelDisplayLabel,
+                model = state.selectedModel,
+                parentMessageId = parentUserMessage.parentMessageId,
+                agentId = if (isAgentContinue) state.selectedModel else null,
+                overrideParentMessageId = parentUserMessage.messageId,
+                responseMessageId = lastAiMessage.messageId,
+                isEdited = true,
+                isRegenerate = true,
+                isContinued = true,
+                webSearch = webSearchEnabled,
+                ephemeralAgent = ephemeralAgent,
+                isTemporary = state.isTemporaryChat,
+            ),
+        )
+    }
+
+    // ── Inline edit-field UI state ───────────────────────────────────
+
+    fun startEditing(messageId: String) {
+        val text = getMessageText(messageId)
+        stateHandle.update { copy(editingMessageId = messageId, editingText = text) }
+    }
+
+    fun onEditTextChanged(text: String) {
+        stateHandle.update { copy(editingText = text) }
+    }
+
+    fun cancelEditing() {
+        stateHandle.update { copy(editingMessageId = null, editingText = "") }
+    }
+
+    fun submitEdit() {
+        val messageId = stateHandle.state.editingMessageId ?: return
+        val newText = stateHandle.state.editingText.trim()
+        if (newText.isBlank()) return
+
+        stateHandle.update { copy(editingMessageId = null, editingText = "") }
+        editMessage(messageId, newText)
+    }
+
+    fun saveEditOnly() {
+        val messageId = stateHandle.state.editingMessageId ?: return
+        val conversationId = stateHandle.state.conversationId ?: return
+        val newText = stateHandle.state.editingText.trim()
+        if (newText.isBlank()) return
+
+        stateHandle.update { copy(editingMessageId = null, editingText = "") }
+        stateHandle.scope.launch {
+            messageRepository.updateMessageText(conversationId, messageId, newText)
+        }
+    }
+}
